scripts/phab2github/requirements.txt
phabricator==0.7.0
gitpython==3.0.6
retrying==1.3.3
